fix: move SecureKeyStore to editor and clean up tests
- split SecureKeyStore into editor-only platform files with CI fallback
- simplify SecureKeyStore tests to skip instead of inconclusive
- restore LazyNodeExecutor cleanup behavior and resolve FILETIME ambiguity
- address nullable warnings in SecureKeyStore.Set
