+++ Unity-MCP-Plugin/Assets/root/Editor/Scripts/Utils/SecureKeyStore.Linux.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afd3729037304e1aacc54fbc0f6ac778
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
