+++ Unity-MCP-Plugin/Assets/root/Editor/Scripts/Utils/SecureKeyStore.Unsupported.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7f3359d87bb40948b5c1c72986b48be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
