+++ Unity-MCP-Plugin/Assets/root/Editor/Scripts/Utils/SecureKeyStore.Windows.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b863d7bc3d9f4f0aa1688b6b94780e00
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
