+++ Unity-MCP-Plugin/Assets/root/Editor/Scripts/Utils/SecureKeyStore.cs.meta
